requirements.txt
paramiko>=4.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
tenacity>=8.2.0
requests>=2.31.0
